requirements.txt
langchain==0.1.7
langgraph
langchain_openai
langchainhub
langsmith
duckduckgo-search
beautifulsoup4
gradio
python-dotenv